requirements.txt
Django==5.2.8
psycopg[binary]==3.2.12
python-dotenv==1.0.1
djangorestframework==3.16.1
django-cors-headers==4.9.0
gunicorn==21.2.0
pytest==9.0.1
pytest-django==4.11.1
django-filter==25.2